Editor asdlaskgh

+++ DBConnect/DBModels/Terrain.cs
@@ -5,7 +5,7 @@
         public int TerrainId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int MovementModifier { get; set; }
+        public decimal MovementModifier { get; set; }
         public bool IsWater { get; set; }
         public bool IsUnderground { get; set; }
         public bool IsRough { get; set; } // Rough terrain cannot be traversed by land vehicles, and typically imposes a movement penalty
